Add files via upload
This is Logistic Regression project to analyse the a dataset of 50,000 IMDB movie reviews as positive and negative. It uses Natural Language Processing to tokenize the documents and also requires pandas library. A review with a rating of atleast 7 is considered to be positive (1) and a review with a rating of atmost 4 is considered to be negative (0).